refactor: 피드백 반영
특별한 이유 없는 static메소드 제거
배열 대신 리스트로 변경
뷰에 의존적인 toString메소드 제거
테스트메소드명 적절하게 변경
ResultTest.java 파일 생성
